+++ tests/TestFabric.Test/TestSuiteNormalTests.cs
@@ -2,6 +2,10 @@
 using JetBrains.Annotations;
 using TestFabric.Data;
 
+// ReSharper disable ReplaceAutoPropertyWithComputedProperty
+
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+
 namespace TestFabric.Test;
 
 [TestSubject(typeof(TestSuite.Normal))]
@@ -447,6 +451,52 @@ public class TestSuiteNormalTests : TestSuite.Normal
         // Assert
         Assert.True(actual.Date >= DateTime.Now.AddDays(-daysBack));
         Assert.True(actual.Date <= DateTime.Now);
+    }
+
+    [Fact]
+    public void Given_DummyClass_When_FromTemplateAge_Then_SameButAgeDifferent()
+    {
+        // Arrange
+        var dummy = Random<DummyClass>();
+
+        // Act
+        var actual = FromTemplate(dummy, x => x.Age);
+
+        // Assert
+        Assert.Equal(dummy.Name, actual.Name);
+        Assert.NotEqual(dummy.Age, actual.Age);
+        Assert.Equal(dummy.Title, actual.Title);
+    }
+
+    [Fact]
+    public void Given_DummyClassAndCount_When_FromTemplateName_Then_MultipleSameButNameDifferent()
+    {
+        // Arrange
+        var dummy = Random<DummyClass>();
+        var count = InRange(3, 7);
+
+        // Act
+        var actual = FromTemplate(
+                count,
+                dummy,
+                x => x.Name)
+            .ToArray();
+
+        // Assert
+        Assert.Equal(count, actual.Length);
+        foreach (var item in actual)
+        {
+            Assert.NotEqual(dummy.Name, item.Name);
+            Assert.Equal(dummy.Age, item.Age);
+            Assert.Equal(dummy.Title, item.Title);
+        }
+    }
+
+    public class DummyClass
+    {
+        public string? Name { get; set; }
+        public int Age { get; set; }
+        public string Title { get; } = "Dr.";
     }
 
     // ReSharper disable once ClassNeverInstantiated.Local