+++ .github/release-notes/release-notes.md
@@ -2,8 +2,7 @@ Provides easier ways to generate randomized data.
 
 ## 🛠 Enhancements
 
-- [#13](https://github.com/zhofre/test-fabric/issues/13) Add float and long pickers.
-- [#13](https://github.com/zhofre/test-fabric/issues/13) Add date pickers.
+- [#11](https://github.com/zhofre/test-fabric/issues/11) Create from template.
 
 ## 📦 NuGet